Added Vector Sprite, Vector Fonts, Fixed Paeth Filter, Added BMP saving, Fixed SVG Filling for Even-Odd rule
