+++ include/PCG.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <iostream>
 
 class PCG
 {
