+++ src/Image_JpegCode.cpp
@@ -1,0 +1,17 @@
+#include "Image.h"
+#include <vector>
+#include "SimpleFile.h"
+#include "Compression.h"
+#include "Graphics.h"
+#include <iostream>
+#include "StringTools.h"
+
+void Image::saveJPG(std::string filename)
+{
+    
+}
+
+Image** Image::loadJPG(std::vector<unsigned char> fileData, int* amountOfImages)
+{
+	return nullptr;
+}
