+++ testFiles/XmlFiles/test2.xml
@@ -1,0 +1,1 @@
+<File>&lt;&#x41;&#x042;&gt;</File>
